feat: Refactor and reorganize components
