+++ src/component/common/helper/apiCall.js
@@ -13,14 +13,14 @@ export const apiCall = async (
     const API_Point = baseURL + url;
 
     const response = await axios({
-      url: API_Point, 
+      url: API_Point,
       method,
       data,
       headers,
       params,
     });
 
-    return response.data; 
+    return response.data;
   } catch (error) {
     console.error("API Call Error:", error);
