+++ src/component/game/commonGameComponent/WinnerHistory/WinnerRectangle.jsx
