+++ src/component/game/commonGameComponent/WinnerHistory/style/WinnerHistory.module.css
