--- src/components/agent/jsx/Header.jsx
@@ -1,6 +1,0 @@
-import React from "react";
-import styles from "../styles/Header.module.css";
-
-export const Header = () => {
-  return <div className={styles.header}>Header</div>;
-};
