--- src/pages/Home.jsx
@@ -1,11 +1,0 @@
-import style from "../styles/Home.module.css";
-
-const Home = () => {
-  return (
-    <div className={style.home}>
-      <h1 className={style.title}>casGamPro</h1>
-    </div>
-  );
-};
-
-export default Home;
