--- src/styles/AdminDashboard.module.css
@@ -1,24 +1,0 @@
-/* src/Page.module.css */
-.pageContainer {
-  display: flex; /* Use Flexbox to arrange items horizontally */
-  color: var(--color-text-primary);
-  height: 100vh; /* Make the page take full height of the viewport */
-}
-
-.sidebar {
-  width: 250px; /* Sidebar width */
-  background-color: var(--color-bg-primary); /* Sidebar background color */
-  color: var(--color-text-primary);
-  padding: 20px;
-  box-shadow: 2px 0 5px rgba(0, 0, 0, 0.1); /* Optional: Add shadow to the sidebar */
-}
-
-.content {
-  flex-grow: 1; /* Content takes the remaining space */
-  padding: 20px;
-  color: var(--color-text-primary);
-  background-color: var(
-    --color-bg-secondary
-  ); /* Content area background color */
-  overflow-y: auto; /* Enable scrolling if the content overflows */
-}
