--- src/styles/AdminPage.css
@@ -1,75 +1,0 @@
-/* Sidebar.css */
-
-/* Sidebar styles */
-.sidebar {
-  width: 250px;
-  height: 100vh;
-  background-color: #333;
-  padding: 20px;
-  color: white;
-  position: fixed;
-  top: 0;
-  left: 0;
-}
-
-/* Dropdown button styles */
-.dropdown {
-  position: relative;
-}
-
-.dropdown-toggle {
-  background-color: #444;
-  border: none;
-  color: white;
-  padding: 10px;
-  width: 100%;
-  text-align: left;
-  font-size: 16px;
-  cursor: pointer;
-  border-radius: 5px;
-}
-
-.dropdown-toggle:hover {
-  background-color: #555;
-}
-
-/* Dropdown menu styles */
-.dropdown-menu {
-  background-color: #555;
-  position: absolute;
-  top: 40px; /* Adjust if needed */
-  width: 100%;
-  border-radius: 5px;
-  padding: 10px 0;
-}
-
-.dropdown-menu ul {
-  list-style-type: none;
-  padding: 0;
-  margin: 0;
-}
-
-.dropdown-menu ul li {
-  padding: 10px;
-  cursor: pointer;
-}
-
-.dropdown-menu ul li:hover {
-  background-color: #666;
-}
-/* Window.css */
-.window {
-  margin-left: 270px; /* Adjust to avoid overlap with sidebar */
-  padding: 20px;
-  background-color: #f4f4f4;
-  height: 100vh;
-}
-
-.window-content {
-  margin-top: 20px;
-}
-
-h2 {
-  font-size: 24px;
-  color: #333;
-}
