--- src/styles/Home.module.css
@@ -1,21 +1,0 @@
-/* Home.css */
-
-/* General styling for the home page */
-.home {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 100vh; /* Full viewport height */
-  margin: 0;
-  background-color: var(
-    --color-bg-primary
-  ); /* Background color based on theme */
-  color: var(--color-text-primary); /* Text color based on theme */
-}
-
-.title {
-  font-size: 4em; /* Large text for the title */
-  font-weight: var(--textBold); /* Bold text */
-  text-align: center;
-  color: var(--color-text-primary); /* Text color based on theme */
-}
