--- src/styles/ManageClient.module.css
@@ -1,30 +1,0 @@
-.newUserBtn {
-  padding: 10px 15px;
-  cursor: pointer;
-  background-color: var(--color-dark-primary); /* Light/dark button color */
-  border-radius: var(--borderRadius); /* Border radius */
-  border: 2px solid var(--color-highlighter-dark-primary); /* Light/dark border color */
-  box-shadow: 6px 6px 10px rgba(0, 0, 0, 1),
-    1px 1px 10px rgba(255, 255, 255, 0.6);
-  color: var(--color-text-primary); /* Light/dark text color */
-  font-size: 18px;
-  font-weight: bold;
-  transition: 0.35s;
-  margin-right: 10px; /* Added margin to the right */
-  float: right; /* Position button to the right side */
-}
-
-.newUserBtn:hover {
-  box-shadow: 6px 6px 10px rgba(0, 0, 0, 1),
-    1px 1px 10px var(--color-highlighter-dark-primary); /* Hover effect */
-}
-
-.newUserBtn:focus {
-  transform: scale(1.05);
-  box-shadow: 6px 6px 10px rgba(0, 0, 0, 1),
-    1px 1px 10px rgba(255, 255, 255, 0.6), inset 2px 2px 10px rgba(0, 0, 0, 1),
-    inset -1px -1px 5px rgba(255, 255, 255, 0.6);
-}
-.button {
-  float: right; /* Float the button to the right */
-}
